Raise warning for potentially malformed student
If you miss the space after the hyphen in them yaml list of students,
it will siltenty assume another list format during unmarshaling and
have each of the items with the hypen (not even sure if this is
correct or should fail instead). Raise a warning as this could be a
mistake.

I don't think there is much common use of either mail, id or username
starting with a hyphen, so I think its fair to raise a warning.